+++ packages/terra-date-picker/tests/wdio/date-picker-spec.js
@@ -72,6 +72,18 @@ describe('Date Picker', () => {
       });
       Terra.should.matchScreenshot('default date cleared');
     });
+
+    describe('Clears default date value that is before the minDate and no maxDate', () => {
+      before(() => browser.url('/#/raw/tests/terra-date-picker/date-picker/date-picker-default-date-before-min-date-only'));
+
+      Terra.should.matchScreenshot('default date displayed');
+      it('clears the default date when focus is on the input', () => {
+        browser.click('input[name="terra-date-date-input"]');
+        // Ensures the mouse pointer doesn't appear in the screenshot
+        browser.click('h3');
+      });
+      Terra.should.matchScreenshot('default date cleared');
+    });
   });
 
   describe('Exclude Dates', () => {